+++ src/lambdaisland/launchpad/env.clj
@@ -83,5 +83,8 @@
 
 (defn watch-handlers []
   (let [h (dotenv-watch-handler [".env" ".env.local"])]
-    {".env" h
-     ".env.local" h}))
+    (cond-> {}
+      (exists? ".env")
+      (assoc ".env" h)
+      (exists? ".env.local")
+      (assoc ".env.local" h))))