+++ CHANGELOG.md
@@ -1,10 +1,8 @@
 # Unreleased
 
-## Added
-
 ## Fixed
 
-## Changed
+- Handle missing `.env` or `.env.local`
 
 # 0.5.33-alpha (2022-09-22 / c37380a)